requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
wordcloud>=1.9.0
openpyxl>=3.1.0
fpdf2>=2.7.0
